Remove CLI help tests
They were useful during Clap 2→3 transition but are more a pain than a help now,
especially with the windows/unix difference. Additionally, the test stack used
is a hinder to going for something a little more modern; the whole CLI testing
approach needs to be rethought for valuable integration testing.

Fixes #362