Add option to exit when stdin ends
The watchexec cli gains a '--stdin-quit' flag which will cause the
watchexec to end & exit when stdin for watchexec closes. This is to make
it easier to use watchexec as a subprocess and make sure that it ends
with the parent process ends, provided the parent process controls the
stdin to the watchexec process.

We model the end of stdin as a keyboard event as it is possible to send
'eof' from a terminal using 'ctrl-d'. We keep reading lines from stdin
and if we get back 0 bytes we assume that means that the stream has
closed. This is a technique I have used elsewhere but I'm not that
experienced so there might be a much better way to do it.