Update keyboard worker to balance config & stdin
We need to watch both the config and stdin and respond to a change in either.
If the config changes, we need to be able to stop watching stdin.

We change to do the stdin watching in a spawned function and pass it a
oneshot channel that we'll use to signal the config changing. The watcher
then monitors both stdin and the oneshot channel and acts appropriately.

If the config changes and we already have a oneshot channel then we know
we must be monitoring stdin already so we take no action. If we don't want
to monitor it and we don't have a channel then we can also take no action.