Use raw_arg for CMD on Windows
CMD uses special handling for arguments to passed to /C. Unfortunately,
this causes errors with quoted program names in these arguments.

This commit implements a workaround: When CMD is requested, we build a
std::process::Command and pass the argument to /C using the special
std::os::windows::process::CommandExt::raw_arg method. The StdCommand is
then converted to a TokioCommand and returned.

Once tokio-rs/tokio#5810 is fixed, this workaround can be removed.

Fixes #613